Allow the response to be more than one element.
Previously we'd grab the "outerHTML" of the first member of the response
element (which is really an array of DOM elements), but that fails if
there's more than one, which can can be the case if the template has
whitespace or comments before the 'html' element.

Instead we put all of the elements into a common container and then
serialize its contents, thus resulting in a complete rendering of the
document.