Make it a fatal error to access location before it's ready.
In practice there's little reason why this should ever happen since we
initialize the location before instantiating anything. However, it can
arise if a module requests $location during its initialization, either
through module.config or module.run. This causes it to be instantiated
during the call to angular.injector, before we've had a chance to set
the request context.

At the moment this just requires modifying the application to not depend
on the location during initialization.