Shim the rendered page so we can still run AngularJS in the browser.
Previously we just ran AngularJS on the server and served completely static
HTML to the browser. This was fine as a prototype of running Angular
on the server, but in real applications we still want to get the full
benefits of AngularJS on the client too, so to move closer to that goal
this shims the static HTML so that, if rendered in a browser that
supports JavaScript, we'll replace the DOM with the template and then load
some other scripts to complete the client-side bootstrap.

At the moment this is only partially implemented in our simple example,
with the shim in place but without a client-side bootstrapping script.
This means that right now the example is actually broken since it'll
just render the template on-screen and never bootstrap in the client.