Make the simple example bootstrap on the client too.
This takes us almost back to having a working app again: the app runs on
the server, renders a static page, and then the shim on the client
injects the original template and loads the app again to get the
client-side version.

However, since the client-side version can't fetch the weather data due
to cross-origin restrictions, this still doesn't work until we get the
server-defined-route facility working and we're thus able to inject
the server's data into the client.