Supply a custom $route implementation, so we can optimize.
The built-in $route doesn't give us any way to resolve a route without
switching to it, so we replace it with a new implementation that behaves
more like what we need for the server-side work.

This allows us to do provide an API to do route resolution on the server.
As of this commit we no longer support rendering a static HTML page,
but that will return in a subsequent commit.