Don't hang if we can't make JSON.
Sometimes the route resolve functions will inadvertently produce something
we can't serialize as JSON, such as a circular data structure, or some
DOM elements. In this case we used to just get stuck, but now we'll
fail hard in that case so the developer can see what went wrong and
fix it.