Return the browser-oriented part of the response sooner.
Browser clients don't actually care about our HTML snapshot, so we can
return the part they do care about as soon as its' ready and then follow
up with the rest of the body later once we have it. In browsers this
will be ignored because we will have replaced the document with the
template by this point.