Forcefully abort pending HTTP requests when disposing.
Our server code (should!) always call context.dispose() at the end of a
request to reclaim the resources used by the context. However, if that
happens when an HTTP request is outstanding, it's important that we
intercept that and prevent any calls back into callbacks that live inside
the context, since those callbacks can end up trying to make use of
objects that are no longer available, which can in turn cause nodejs to
crash.