Override the internal $$SanitizeUri service to speed it up.
The real target here is the internal urlResolve function, but since we
can't easily override that (it's just a local function, not a service)
we do the next best thing and override the most common caller of that
function instead. This new implementation uses the native nodejs URL
resolving function, rather than the hack of assigning a URL to the href
attribute of a detached link tag.

The downside of this change is that it overrides an interface that is
considered internal to AngularJS, and thus might change in future versions.
In that case we'll need to find a better way to solve this problem.