Fail gracefully if a caller attempts an HTTP request before we're ready.
We can't do an HTTP request until we know our URL. Before this was a
fatal error that would cause a request to get "stuck" in the queue,
making the "no more requests" event never fire. After this change it's
still an error, but one surfaced through the callback interface so that
the caller might actually handle it, and so that we can count the
request as 'resolved'.

This situation unfortunately tends to arise if there's an ng-include in
the template, because we compile the template before we know our URL.
Living with that issue for now, but we'll probably have to come back and
fix it properly at some point.