Override $sceDelegate to be a no-op.
This is a performance optimization since SCE is a very common caller to
urlResolve. We can't easily override urlResolve since it's not part of
a service, but by overriding some of its most common callers we can get
much of the benefit of doing so.

In this case rather than trying to provide a server-specific implementation
of the service we just make it a no-op. We might want to revisit this
later to implement some rules that make sense in the server context, but
for now we assume that the attack surface of our server is much smaller
than a browser (since it doesn't execute user-provided scripts, doesn't
render links to the user, etc.) and just take the performance benefit of
not running this code at all.