Remove redundant route resolution functions.
An earlier design for this library had these functions as methods on
the server object, but they've since moved to be methods on the $route
service and its route instances. The SDR API and HTML snapshot helpers
got left behind on the old implementations, causing some unnecessary
slight differences in behavior.