Defer compiling the template until HTML generation.
Users of runInContext or wrapMiddlewareWithAngular may be performing
operations that don't require a template to be compiled at all, so we
skip that step by default but opt into it for the HTML generator, which
*does* need it. Other callers that need a root element can similarly
call injector.bootstrap with an element to use as $rootElement for the
injector.

The other attractive side-effect of this change is that we defer template
compilation until we've associated the injector with a request, which
means that directives are guaranteed to be able to use $location during
compilation, if necessary.