Beginning of a revamp of the weather example.
Started to spruce it up a bit to make it a bit prettier, both in terms
of visual style (now using Bootstrap CSS) and in terms of code (now able
to use templateUrl to reference templates rather than inlining them
in the JS.)

More work is required here. In particular, this has broken the index page.
This is just a checkpoint.